requirements.txt
flask
pymysql
Flask-SQLAlchem